+++ lib/controller.js
@@ -3,7 +3,6 @@ const Zigbee = require('./zigbee');
 const logger = require('./util/logger');
 const settings = require('./util/settings');
 const deviceMapping = require('./devices');
-const mqtt2zigbee = require('./converters/mqtt2zigbee');
 const homeassistant = require('./homeassistant');
 const debug = require('debug')('zigbee2mqtt:controller');
 
@@ -295,13 +294,14 @@ class Controller {
 
         Object.keys(json).forEach((key) => {
             // Find converter for this key.
-            const converter = mqtt2zigbee[key];
+            const converter = mappedModel.toZigbee.find((c) => c.key === key);
+
             if (!converter) {
                 logger.error(`No converter available for '${key}' (${json[key]})`);
                 return;
             }
 
-            const message = converter(json[key]);
+            const message = converter.convert(json[key]);
             const callback = (error) => {
                 // Devices do not report when they go off, this ensures state (on/off) is always in sync.
                 if (!error && key.startsWith('state')) {
