+++ lib/devices.js
@@ -1,4 +1,5 @@
 const fz = require('./converters/fromZigbee');
+const tz = require('./converters/toZigbee');
 
 const LED1623G12 = {
     model: 'LED1623G12',
@@ -6,6 +7,7 @@ const LED1623G12 = {
     description: 'TRADFRI LED bulb E27 1000 lumen, dimmable, opal white',
     supports: 'on/off, brightness',
     fromZigbee: [fz.light_brightness, fz.ignore_onoff_change],
+    toZigbee: [tz.onoff, tz.light_brightness],
 };
 
 const LED1536G5 = {
@@ -14,6 +16,7 @@ const LED1536G5 = {
     description: 'TRADFRI LED bulb E12/E14 400 lumen, dimmable, white spectrum, opal white',
     supports: 'on/off, brightness, color temperature',
     fromZigbee: [fz.light_brightness, fz.light_color_colortemp, fz.ignore_onoff_change],
+    toZigbee: [tz.onoff, tz.light_brightness, tz.light_colortemp],
 };
 
 const devices = {
@@ -24,6 +27,7 @@ const devices = {
         description: 'MiJia wireless switch',
         supports: 'single, double, triple, quadruple, many and long click',
         fromZigbee: [fz.xiaomi_battery_3v, fz.WXKG01LM_click, fz.ignore_onoff_change, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.sensor_switch.aq2': {
         model: 'WXKG11LM',
@@ -31,6 +35,7 @@ const devices = {
         description: 'Aqara wireless switch',
         supports: 'single, double, triple, quadruple click',
         fromZigbee: [fz.xiaomi_battery_3v, fz.WXKG11LM_click, fz.ignore_onoff_change, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.sensor_86sw1\u0000lu': {
         model: 'WXKG03LM',
@@ -38,6 +43,7 @@ const devices = {
         description: 'Aqara single key wireless wall switch',
         supports: 'single click',
         fromZigbee: [fz.xiaomi_battery_3v, fz.WXKG03LM_click],
+        toZigbee: [],
     },
     'lumi.sensor_86sw2\u0000Un': {
         model: 'WXKG02LM',
@@ -45,6 +51,7 @@ const devices = {
         description: 'Aqara double key wireless wall switch',
         supports: 'left, right and both click',
         fromZigbee: [fz.xiaomi_battery_3v, fz.WXKG02LM_click],
+        toZigbee: [],
     },
     'lumi.ctrl_neutral1': {
         model: 'QBKG04LM',
@@ -52,6 +59,7 @@ const devices = {
         description: 'Aqara single key wired wall switch',
         supports: 'on/off',
         fromZigbee: [fz.QBKG04LM_state, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff],
         ep: {'': 2},
     },
     'lumi.ctrl_neutral2': {
@@ -60,6 +68,7 @@ const devices = {
         description: 'Aqara double key wired wall switch',
         supports: 'l1 and l2 on/off',
         fromZigbee: [fz.QBKG03LM_state, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff],
         ep: {'l1': 2, 'l2': 3},
     },
     'lumi.sens': {
@@ -68,6 +77,7 @@ const devices = {
         description: 'MiJia temperature & humidity sensor ',
         supports: 'temperature and humidity',
         fromZigbee: [fz.xiaomi_battery_3v, fz.xiaomi_temperature, fz.xiaomi_humidity, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.weather': {
         model: 'WSDCGQ11LM',
@@ -78,6 +88,7 @@ const devices = {
             fz.xiaomi_battery_3v, fz.xiaomi_temperature, fz.xiaomi_humidity, fz.xiaomi_pressure, fz.ignore_basic_change,
             fz.ignore_temperature_change, fz.ignore_humidity_change, fz.ignore_pressure_change,
         ],
+        toZigbee: [],
     },
     'lumi.sensor_motion': {
         model: 'RTCGQ01LM',
@@ -85,6 +96,7 @@ const devices = {
         description: 'MiJia human body movement sensor',
         supports: 'occupancy',
         fromZigbee: [fz.xiaomi_battery_3v, fz.xiaomi_occupancy, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.sensor_motion.aq2': {
         model: 'RTCGQ11LM',
@@ -95,6 +107,7 @@ const devices = {
             fz.xiaomi_battery_3v, fz.xiaomi_occupancy, fz.xiaomi_illuminance, fz.ignore_basic_change,
             fz.ignore_illuminance_change, fz.ignore_occupancy_change,
         ],
+        toZigbee: [],
     },
     'lumi.sensor_magnet': {
         model: 'MCCGQ01LM',
@@ -102,6 +115,7 @@ const devices = {
         description: 'MiJia door & window contact sensor',
         supports: 'contact',
         fromZigbee: [fz.xiaomi_battery_3v, fz.xiaomi_contact, fz.ignore_onoff_change, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.sensor_magnet.aq2': {
         model: 'MCCGQ11LM',
@@ -109,6 +123,7 @@ const devices = {
         description: 'Aqara door & window contact sensor',
         supports: 'contact',
         fromZigbee: [fz.xiaomi_battery_3v, fz.xiaomi_contact, fz.ignore_onoff_change, fz.ignore_basic_change],
+        toZigbee: [],
     },
     'lumi.sensor_wleak.aq1': {
         model: 'SJCGQ11LM',
@@ -119,6 +134,7 @@ const devices = {
             fz.xiaomi_battery_3v, fz.SJCGQ11LM_water_leak_basic, fz.SJCGQ11LM_water_leak_iaszone,
             fz.ignore_basic_change,
         ],
+        toZigbee: [],
     },
     'lumi.sensor_cube': {
         model: 'MFKZQ01LM',
@@ -129,6 +145,7 @@ const devices = {
             fz.xiaomi_battery_3v, fz.MFKZQ01LM_action_multistate, fz.MFKZQ01LM_action_analog,
             fz.ignore_analog_change, fz.ignore_multistate_change,
         ],
+        toZigbee: [],
     },
     'lumi.plug': {
         model: 'ZNCZ02LM',
@@ -139,6 +156,7 @@ const devices = {
             fz.xiaomi_state, fz.xiaomi_power, fz.ZNCZ02LM_state, fz.ignore_onoff_change,
             fz.ignore_basic_change, fz.ignore_analog_change,
         ],
+        toZigbee: [tz.onoff],
     },
     'lumi.ctrl_86plug': {
         model: 'QBCZ11LM',
@@ -146,6 +164,7 @@ const devices = {
         supports: 'on/off, power measurement',
         vendor: 'Xiaomi',
         fromZigbee: [fz.xiaomi_state, fz.xiaomi_power, fz.ignore_onoff_change, fz.ignore_analog_change],
+        toZigbee: [tz.onoff],
     },
     'lumi.sensor_smoke': {
         model: 'JTYJ-GD-01LM/BW',
@@ -153,6 +172,7 @@ const devices = {
         supports: 'smoke',
         vendor: 'Xiaomi',
         fromZigbee: [fz.xiaomi_battery_3v, fz.JTYJGD01LMBW_smoke, fz.ignore_basic_change],
+        toZigbee: [],
     },
 
     // IKEA
@@ -162,6 +182,7 @@ const devices = {
         description: 'TRADFRI LED bulb E27 980 lumen, dimmable, white spectrum, opal white',
         supports: 'on/off, brightness, color temperature',
         fromZigbee: [fz.light_brightness, fz.light_color_colortemp, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness, tz.light_colortemp],
     },
     // LED1623G12 uses 2 model IDs, see https://github.com/Koenkk/zigbee2mqtt/issues/21
     'TRADFRI bulb E27 opal 1000lm': LED1623G12,
@@ -172,6 +193,7 @@ const devices = {
         description: 'TRADFRI LED bulb GU10 400 lumen, dimmable, white spectrum',
         supports: 'on/off, brightness, color temperature',
         fromZigbee: [fz.light_brightness, fz.light_color_colortemp, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness, tz.light_colortemp],
     },
     'TRADFRI bulb GU10 W 400lm': {
         model: 'LED1650R5',
@@ -179,6 +201,7 @@ const devices = {
         description: 'TRADFRI LED bulb GU10 400 lumen, dimmable',
         supports: 'on/off, brightness',
         fromZigbee: [fz.light_brightness, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness],
     },
     // LED1536G5 has an E12 and E14 version.
     'TRADFRI bulb E14 WS opal 400lm': LED1536G5,
@@ -189,6 +212,7 @@ const devices = {
         description: 'TRADFRI LED bulb E26 1000 lumen, dimmable, opal white',
         supports: 'on/off, brightness',
         fromZigbee: [fz.light_brightness, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness],
     },
 
     // Philips
@@ -198,6 +222,7 @@ const devices = {
         description: 'Hue Go',
         supports: 'on/off, brightness, color temperature, color xy',
         fromZigbee: [fz.light_brightness, fz.light_color_colortemp, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness, tz.light_colortemp, tz.light_color],
     },
 
     // Belkin
@@ -207,6 +232,7 @@ const devices = {
         description: 'WeMo smart LED bulb',
         supports: 'on/off, brightness',
         fromZigbee: [fz.light_brightness, fz.ignore_onoff_change],
+        toZigbee: [tz.onoff, tz.light_brightness],
     },
 
     // EDP
@@ -216,6 +242,7 @@ const devices = {
         description: 're:dy plug',
         supports: 'on/off, power measurement',
         fromZigbee: [fz.ignore_onoff_change, fz.EDP_power, fz.ignore_metering_change],
+        toZigbee: [tz.onoff],
         report: [{
             'cid': 'seMetering',
             'attr': 'instantaneousDemand',
@@ -233,6 +260,7 @@ const devices = {
         description: '[CC2530 router](http://ptvo.info/cc2530-based-zigbee-coordinator-and-router-112/)',
         supports: 'state, description, type, rssi',
         fromZigbee: [fz.CC2530ROUTER_state, fz.CC2530ROUTER_meta],
+        toZigbee: [],
     },
 };
